LayoutTests/media/session/switching-sessions-expected.txt

When switching media sessions, media elements belonging to the old session should be paused.


RUN(session = new MediaSession("content"))
EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
Begin playing media.
EVENT(playing)
Media began playing.
EXPECTED (video.paused == 'false') OK
RUN(video.session = new MediaSession("transient"))
EXPECTED (video.paused == 'true') OK
END OF TEST
